# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Start the AutoML pipeline for a given Kaggle dataset
 * @summary Run AutoML pipeline
 */
export const RunPipelineBody = zod.object({
  datasetName: zod
    .string()
    .describe(
      'Kaggle dataset identifier (e.g. \"titanic\" or \"username\/dataset-name\")',
    ),
  targetColumn: zod
    .string()
    .nullish()
    .describe("Optional target column name. Auto-detected if not provided."),
  kaggleUsername: zod
    .string()
    .nullish()
    .describe("Optional Kaggle username for API auth"),
  kaggleKey: zod
    .string()
    .nullish()
    .describe("Optional Kaggle API key for auth"),
});

/**
 * @summary Get pipeline job status
 */
export const GetPipelineStatusParams = zod.object({
  jobId: zod.coerce.string(),
});

export const GetPipelineStatusResponse = zod.object({
  jobId: zod.string(),
  status: zod.enum(["pending", "running", "completed", "failed"]),
  stage: zod.string().nullish().describe("Current stage description"),
  datasetName: zod.string(),
  targetColumn: zod.string().nullish(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
  error: zod.string().nullish(),
});

/**
 * @summary Get pipeline job result
 */
export const GetPipelineResultParams = zod.object({
  jobId: zod.coerce.string(),
});

export const GetPipelineResultResponse = zod.object({
  jobId: zod.string(),
  datasetName: zod.string(),
  targetColumn: zod.string(),
  problemType: zod.enum(["classification", "regression"]),
  bestModel: zod.string(),
  bestScore: zod.number(),
  scoreMetric: zod.string(),
  modelScores: zod.array(
    zod.object({
      modelName: zod.string(),
      score: zod.number(),
      trainingTime: zod.number(),
    }),
  ),
  featureImportance: zod.array(
    zod.object({
      feature: zod.string(),
      importance: zod.number(),
    }),
  ),
  featureNames: zod
    .array(zod.string())
    .describe(
      "Feature names used during training (for building prediction form)",
    ),
  datasetSummary: zod.object({
    rows: zod.number(),
    columns: zod.number(),
    columnNames: zod.array(zod.string()),
    missingValues: zod.number(),
    duplicatesRemoved: zod.number(),
    categoricalColumns: zod.array(zod.string()),
    numericColumns: zod.array(zod.string()),
  }),
  modelPath: zod.string().nullish(),
});

/**
 * @summary List recent pipeline jobs
 */
export const ListPipelineJobsResponseItem = zod.object({
  jobId: zod.string(),
  status: zod.enum(["pending", "running", "completed", "failed"]),
  stage: zod.string().nullish().describe("Current stage description"),
  datasetName: zod.string(),
  targetColumn: zod.string().nullish(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
  error: zod.string().nullish(),
});
export const ListPipelineJobsResponse = zod.array(ListPipelineJobsResponseItem);

/**
 * @summary Run prediction on trained model
 */
export const PredictPipelineParams = zod.object({
  jobId: zod.coerce.string(),
});

export const PredictPipelineBody = zod.object({
  features: zod
    .record(zod.string(), zod.string())
    .describe("Map of feature name to string value entered by user"),
});

export const PredictPipelineResponse = zod.object({
  predictedValue: zod
    .string()
    .describe("The model's predicted output as a string"),
  topFeatures: zod
    .array(
      zod.object({
        feature: zod.string(),
        importance: zod.number(),
      }),
    )
    .describe("Top 5 most important features"),
  explanation: zod
    .string()
    .describe("Human-readable explanation of prediction"),
  confidence: zod
    .string()
    .nullish()
    .describe("High, Medium, or Low confidence level"),
  confidenceDetail: zod
    .string()
    .nullish()
    .describe("Metric value used to compute confidence"),
  problemType: zod.enum(["classification", "regression"]),
});

/**
 * @summary Search Kaggle datasets
 */
export const SearchKaggleDatasetsQueryParams = zod.object({
  query: zod.coerce.string(),
});

export const SearchKaggleDatasetsResponseItem = zod.object({
  ref: zod.string(),
  title: zod.string(),
  size: zod.string().nullish(),
  lastUpdated: zod.string().nullish(),
  downloadCount: zod.number().nullish(),
});
export const SearchKaggleDatasetsResponse = zod.array(
  SearchKaggleDatasetsResponseItem,
);